requirements.txt
# LLM API clients
anthropic>=0.18.0
openai>=1.12.0

# Embeddings and NLP
sentence-transformers>=2.2.0
transformers>=4.36.0

# Numerical operations
numpy>=1.24.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0

# Development tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
